change convertor & add textNode js
